Ported to .NET Core & created web app!
Port was very easy (obviously). The core project takes in a string instead of a file path. The web app works (really?!) but is full of bodges and is awful to look at. But it works. Will replace the old compiler project with the new core one with windows-only UI probably. Embedded language files into the core project and linked to web project with dll.